+++ vm/src/deal_id.rs
@@ -2,4 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 /// Deal identifier used in market and miner actors
-pub type DealID = u64;
+pub type DealID = usize;
